+++ daas/daas_app/tests/unit_tests/test_task_status.py
@@ -7,6 +7,8 @@ from ...models import Sample, Task, Result
 from ...utils.task_manager import TaskManager
 from ..test_utils.resource_directories import CSHARP_SAMPLE, TEXT_SAMPLE, CSHARP_BROKEN
 
+# Test run on different classes to avoid port overlapping between tests and guarantee test independence.
+
 
 class JobStatusTestCase(NonTransactionalLiveServerTestCase):
     def _wipe(self) -> None:
